SONARPY-186 Rule: Cognitive Complexity of functions should not be too high

+++ python-checks/src/test/java/org/sonar/python/checks/utils/PythonCheckVerifier.java
@@ -80,7 +80,7 @@ public class PythonCheckVerifier extends SquidCheck<Grammar> implements AstAndTo
       assertThat(actual.primaryLocation().message()).as("Bad message at line " + expected.line()).isEqualTo(expected.message());
     }
     if (expected.effortToFix() != null) {
-      assertThat(actual.cost()).as("Bad effortToFix at line " + expected.line()).isEqualTo(expected.effortToFix());
+      assertThat(actual.cost().intValue()).as("Bad effortToFix at line " + expected.line()).isEqualTo(expected.effortToFix());
     }
     if (expected.startColumn() != null) {
       assertThat(actual.primaryLocation().startLineOffset() + 1).as("Bad start column at line " + expected.line()).isEqualTo(expected.startColumn());
